目前已经能正常运行了，就是还没做太多测试

+++ src/models/tables.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from enum import Enum
 
-from sqlalchemy import String, Integer, DateTime, ForeignKey, Float, BigInteger, Index
+from sqlalchemy import String, DateTime, ForeignKey, Float, Integer, Index
 from sqlalchemy.orm import mapped_column, Mapped, relationship
 
 from models.base import Base
@@ -18,7 +18,7 @@ class Match(Base):
     """比赛记录"""
     __tablename__ = 'match'
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     choice: Mapped[str] = mapped_column(String(100), index=True, comment='我选的队伍')
     winner: Mapped[str | None] = mapped_column(String(100), nullable=True, comment='胜者，输了或平局时为空')
     result: Mapped[str | None] = mapped_column(String(100), nullable=True, comment='"win"/"lose"/"draw"/None')
@@ -42,8 +42,8 @@ class MatchTeam(Base):
     """比赛参赛队伍"""
     __tablename__ = 'match_team'
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    match_id: Mapped[int] = mapped_column(BigInteger, ForeignKey('match.id', ondelete='CASCADE'))
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    match_id: Mapped[int] = mapped_column(Integer, ForeignKey('match.id', ondelete='CASCADE'))
     name: Mapped[str] = mapped_column(String(100), index=True)
     odds: Mapped[float | None] = mapped_column(Float, nullable=True, comment='倍率')
 
@@ -60,8 +60,8 @@ class BattleRecord(Base):
     __table_args__ = (
         Index('idx_battle_winner_loser', 'winner', 'loser'),
     )
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    match_id: Mapped[int] = mapped_column(BigInteger, ForeignKey('match.id', ondelete='CASCADE'))
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    match_id: Mapped[int] = mapped_column(Integer, ForeignKey('match.id', ondelete='CASCADE'))
     winner: Mapped[str] = mapped_column(String(100), index=True)
     loser: Mapped[str] = mapped_column(String(100), index=True)
     create_time: Mapped[datetime] = mapped_column(DateTime, default=datetime.now)
